+++ src/ModelGenerator.Tests/CSharp/CSharpGeneratorTests.cs
@@ -26,7 +26,7 @@
           }
         };
       var specAnalyzer = new SpecAnalyzer(spec);
-      var generator = new CSharpGenerator(specAnalyzer, new CSharpEntityGeneratorFactory(new CSharpEntityMemberGeneratorFactory()));
+      var generator = new CSharpGenerator(specAnalyzer);
 
       // act
       var outputs = generator.GenerateOutputs();
@@ -56,7 +56,7 @@
         };
 
       var specAnalyzer = new SpecAnalyzer(spec);
-      var generator = new CSharpGenerator(specAnalyzer, new CSharpEntityGeneratorFactory(new CSharpEntityMemberGeneratorFactory()));
+      var generator = new CSharpGenerator(specAnalyzer);
 
       // act
       var outputs = generator.GenerateOutputs();
@@ -108,7 +108,7 @@
       var ammendment = new AmmendmentFactory().CreateAmmendment(Constants.CSharpTarget);
       ammendment.AmmedSpecification(spec);
       var specAnalyzer = new SpecAnalyzer(spec);
-      var generator = new CSharpGenerator(specAnalyzer, new CSharpEntityGeneratorFactory(new CSharpEntityMemberGeneratorFactory()));
+      var generator = new CSharpGenerator(specAnalyzer);
 
       // act
       var outputs = generator.GenerateOutputs();
@@ -227,7 +227,7 @@
       var ammendment = new AmmendmentFactory().CreateAmmendment(Constants.CSharpTarget);
       ammendment.AmmedSpecification(spec);
       var specAnalyzer = new SpecAnalyzer(spec);
-      var generator = new CSharpGenerator(specAnalyzer, new CSharpEntityGeneratorFactory(new CSharpEntityMemberGeneratorFactory()));
+      var generator = new CSharpGenerator(specAnalyzer);
 
       // act
       var outputs = generator.GenerateOutputs();
@@ -279,7 +279,7 @@
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].Type == "string" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.IssueRequiredAllowEmptyStrings &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].Name == "Posts" &&
-          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].Type == "System.Collections.Generic.IList<BlogPost>" &&
+          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].Type == "IList<BlogPost>" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.IssueRequired &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].Name == "Author" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].Type == "Author" &&
@@ -296,13 +296,13 @@
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[0].Type == "int" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[0].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.NoRequiredAttribute &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].Name == "DatePublished" &&
-          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].Type == "System.DateTimeOffset?" &&
+          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].Type == "DateTimeOffset?" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[1].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.NoRequiredAttribute &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].Name == "Description" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].Type == "string" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[2].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.NoRequiredAttribute &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].Name == "Comments" &&
-          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].Type == "System.Collections.Generic.IList<Comment>" &&
+          ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].Type == "IList<Comment>" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[3].RequiredAttributeBehavior == CSharpRequiredAttributeBehavior.IssueRequired &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[4].Name == "Status" &&
           ((CSharpClass)((CSharpNamespace)_.GenerationRoot).Types[0]).Members[4].Type == "BlogPostStatus" &&